Bugfix for non-internet mode
Without Google key & engine ID defined, a task sometimes gets rephrased to an internet search request. This happens when the LLM decides that an internet search would be good, but does not get a reply from the respective API in file google_search.py.

I did fix this. But for the time being w/o internet search there is no information about the need for a search in the output.